# apps/api-go/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Album:
    properties:
      body:
        example: <p>Hello world</p>
        type: string
      created_at:
        example: "2019-04-19T17:47:28Z"
        type: string
      id:
        example: 1
        type: integer
      meta_description:
        example: A good meta
        type: string
      notify_users_on_published:
        example: true
        type: boolean
      private:
        example: true
        type: boolean
      published_at:
        example: "2019-04-19T17:47:28Z"
        type: string
      slug:
        example: a-good-album
        type: string
      sso_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      title:
        example: A good album
        type: string
      updated_at:
        example: "2019-04-19T17:47:28Z"
        type: string
      user_id:
        example: 1
        type: number
    type: object
host: localhost:1323
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /albums:
    get:
      consumes:
      - application/json
      description: Get paginated list of albums
      parameters:
      - default: 1
        description: page
        in: path
        name: page
        required: true
        type: integer
      - default: 10
        description: per_page
        in: path
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Album'
            type: array
      summary: List albums
  /albums/{slug}:
    get:
      consumes:
      - application/json
      description: Get an allbum by slug
      parameters:
      - description: Album Slug
        in: path
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Album'
      summary: Show a album
swagger: "2.0"